requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
blas=1.0=mkl
ca-certificates=2021.10.26=h06a4308_2
certifi=2016.2.28=py36_0
cudatoolkit=9.2=0
cycler=0.10.0=py36_0
dataclasses=0.8=pyh4f3eec9_6
dbus=1.13.18=hb2f20db_0
expat=2.4.1=h2531618_2
fontconfig=2.13.1=h6c09931_0
freetype=2.11.0=h70c0345_0
glib=2.56.2=hd408876_0
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=hb453b48_1
icu=58.2=he6710b0_3
imageio=2.9.0=pyhd3eb1b0_0
intel-openmp=2021.4.0=h06a4308_3561
jbig=2.1=0
jpeg=9b=0
kiwisolver=1.3.1=py36h2531618_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.35.1=h7274673_9
libffi=3.2.1=1
libgcc=5.2.0=0
libgcc-ng=9.3.0=h5101ec6_17
libgfortran=3.0.0=1
libgomp=9.3.0=h5101ec6_17
libiconv=1.14=0
libpng=1.6.37=hbc83047_0
libprotobuf=3.4.0=0
libstdcxx-ng=9.3.0=hd4cf53a_17
libtiff=4.2.0=h85742a9_0
libuuid=1.0.3=0
libuv=1.40.0=h7b6447c_0
libwebp-base=1.2.0=h27cfd23_0
libxcb=1.14=h7b6447c_0
libxml2=2.9.12=h03d6c58_0
lz4-c=1.9.3=h295c915_1
matplotlib=2.2.2=py36hb69df0a_2
mkl=2019.4=243
mkl-service=2.3.0=py36he8ac12f_0
mpmath=0.19=py36_1
ncurses=6.3=h7f8727e_2
ninja=1.7.2=0
nomkl=2.0=0
numpy=1.13.1=py36_nomkl_0
numpy-base=1.19.2=py36hfa32c7d_0
olefile=0.44=py36_0
openblas=0.2.19=0
openssl=1.1.1l=h7f8727e_0
pcre=8.45=h295c915_0
pillow=7.2.0=py36hb39fc2d_0
pip=9.0.1=py36_1
protobuf=3.19.1=pypi_0
pydensecrf=1.0rc2=py36_0
pyparsing=2.2.0=py36_0
pyqt=5.6.0=py36_2
python=3.6.10=hcf32534_1
python-dateutil=2.6.1=py36_0
pytorch=1.7.0=py3.6_cuda9.2.148_cudnn7.6.3_0
pytz=2017.2=py36_0
qt=5.6.3=h8bf5577_3
randaugment=1.0.2=pypi_0
readline=8.1=h27cfd23_0
scipy=0.19.1=np113py36_nomkl_0
setuptools=36.4.0=py36_1
sip=4.18=py36_0
six=1.10.0=py36_0
sqlite=3.36.0=hc218d9a_0
sympy=1.1.1=py36_0
tensorboardx=2.2=pyhd3eb1b0_0
tk=8.6.11=h1ccaba5_0
torchaudio=0.7.0=py36
torchvision=0.8.0=py36_cu92
tornado=4.5.2=py36_0
tqdm=4.15.0=py36_0
typing_extensions=3.10.0.2=pyh06a4308_0
wheel=0.29.0=py36_0
xz=5.2.5=h7b6447c_0
zlib=1.2.11=0
zstd=1.4.5=h9ceee32_0
